pyproject.toml
[project]
name = "ocr-donut"
version = "0.1.0"
requires-python = ">=3.10"
dependencies = [
    "transformers>=4.38.2",
    "torch>=2.2.0; platform_system == 'Darwin' and platform_machine == 'arm64'",
    "pdf2image",
    "pillow",
    "python-dotenv"
]

[tool.black]
line-length = 88
target-version = ["py310"]
include = '\.pyi?$'

[tool.isort]
profile = "black"
multi_line_output = 3
line_length = 88

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
check_untyped_defs = true

[tool.ruff]
select = ["E", "F", "B", "W", "I", "N", "UP", "ANN", "S", "BLE"]
ignore = ["ANN101"]
target-version = "py310"
line-length = 88

[project.scripts]
ocr-llm = "src.interfaces.cli:main"

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
addopts = "-v --cov=src"

[tool.coverage.run]
source = ["src"]